Improve subset retrieval
- implement random and regular image sampling

- fix retrieval based on location tags

- solve overwriting of existing subset files and directories

+++ img_db.py
@@ -33,9 +33,11 @@ import datetime
 import csv
 import hashlib
 import argparse
+import copy
 
 import numpy as np
 
+import utils
 
 parser = argparse.ArgumentParser(description="Query image database.")
 parser.add_argument(
@@ -82,6 +84,32 @@ parser.add_argument(
     choices=["relpath_to_name", "name", "keep_relpaths"], 
     help="The mode of the naming of the copied images."
 )
+parser.add_argument(
+    "--append_mode",
+    type=str,
+    default="overwrite",
+    choices=["keep_new", "keep_old", "overwrite"],
+    help="The mode of the file saving if the there are already existing files in the given paths."
+)
+parser.add_argument(
+    "--subset_size",
+    type=int,
+    default=-1,
+    help="Limit the number of images in the filtered set."
+)
+parser.add_argument(
+    "--subset_mode",
+    type=str,
+    default="random",
+    choices=["random", "regular"],
+    help="The mode of the subset selection - sample randomly or regularly."
+)
+parser.add_argument(
+    "--random_seed",
+    type=int,
+    default=42,
+    help="The seed for the random subset selection."
+)
 
 
 def main(args):
@@ -90,14 +118,17 @@ def main(args):
     db_orig.load(args.db_path)
     db_filt = db_orig.filter(args.query)
 
+    if args.subset_size != -1:
+        db_filt = db_filt.get_subset(args.subset_size, mode=args.subset_mode, seed=args.random_seed)
+
     if args.db_out is not None:
-        db_filt.save(args.db_out)
+        db_filt.save(args.db_out, args.append_mode)
     
     if args.img_list is not None:
-        db_filt.create_image_list_file(args.img_list, mode=args.img_list_mode)
+        db_filt.create_image_list_file(args.img_list, img_list_mode=args.img_list_mode, append_mode=args.append_mode)
     
     if args.img_dir is not None:
-        db_filt.copy_images(args.img_dir, mode=args.img_dir_mode)
+        db_filt.copy_images(args.img_dir, mode=args.img_dir_mode, append_mode=args.append_mode)
 
 
 class ImageDatabase:
@@ -258,7 +289,7 @@ class ImageDatabase:
                 date = datetime.datetime.strptime(row[1] + " " + row[2] + " " + row[3], "%Y-%m-%d %H:%M:%S.%f %z")
                 sensor_name = row[4]
                 sequence = row[5]
-                tag_location = row[6]
+                tag_location = str2list(row[6])
                 tag_daytime = row[7]
                 if row[8] == "-":
                     coords_wgs84 = None
@@ -287,28 +318,43 @@ class ImageDatabase:
                 }
 
 
-    def save(self, image_database_path):
+    def save(self, image_database_path, append_mode="overwrite"):
         """Save the image database to a file.
 
         Parameters:
         image_database_path (str): The path to the image database file.
+        keep (str): The mode of the file saving if the database file already 
+        exists. Can be "keep_new" (keep the new data in case of overlap), 
+        "keep_old" (keep the old data in case of overlap), or "overwrite" 
+        (overwrite all of the old data).
 
         """
 
         if os.path.isfile(image_database_path):
-            file_mode = "at"
+            print("- the given image database file already exists")
+            if append_mode == "keep_new":
+                print("  - keeping the new image data in case of overlap")
+                save_db = merge_databases(image_database_path, self, append_mode)
+            elif append_mode == "keep_old":
+                print("  - keeping the old image data in case of overlap")
+                save_db = merge_databases(image_database_path, self, append_mode)
+            elif append_mode == "overwrite":
+                print("  - overwriting the existing image database file")
+                save_db = self
+                
         else:
-            file_mode = "wt"
+            print("- creating a new image database file")
+            save_db = self
 
-        with open(image_database_path, mode=file_mode, newline='') as f:
+        with open(image_database_path, mode="wt", newline='') as f:
             csv_writer = csv.writer(f)
 
-            image_relpath_list = list(self.db.keys())
+            image_relpath_list = list(save_db.db.keys())
             image_relpath_list.sort()
 
             for img_relpath in image_relpath_list:
                 img_name = os.path.basename(img_relpath)
-                img_data = self.db[img_relpath]
+                img_data = save_db.db[img_relpath]
                 date = img_data["capture_time"].strftime("%Y-%m-%d")
                 time = img_data["capture_time"].strftime("%H:%M:%S.%f")[:-3]
                 timezone = img_data["capture_time"].strftime("%z")
@@ -386,6 +432,10 @@ class ImageDatabase:
 
         """
 
+        filter_str = filter_str.strip()
+        if len(filter_str) == 0:
+            return copy.deepcopy(self)
+
         filter_str = filter_str.lower()
         filter_dict = {}
         for substr in filter_str.split(";"):
@@ -419,6 +469,7 @@ class ImageDatabase:
             filter_dict[key] = {"mode": mode, "values": filter_values}
 
         filtered_db = ImageDatabase(root_path=self.root_path)
+        passed_nums = {key: 0 for key in filter_dict}
 
         for img_name, img_data in self.db.items():
             passed = True
@@ -489,18 +540,23 @@ class ImageDatabase:
                 
                 cyclic_value = key in ["capture_hour", "capture_month", "capture_year", "longitude"]
 
-                passed = passed and compare_metadata(image_value, filter_values, filter_mode, cyclic=cyclic_value)
-
-                if not passed:
-                    break
+                passed_this = compare_metadata(image_value, filter_values, filter_mode, cyclic=cyclic_value)
+                passed_nums[key] += passed_this
+                passed = passed and passed_this
 
             if passed:
                 filtered_db.add_image(img_name, img_data)
 
+        print("- number of images within each category:")
+        for key, num in passed_nums.items():
+            print("  - {}: {} images".format(key, num))
+        print("  ----------------------------------------")
+        print("  - total retrieved images: {}".format(len(filtered_db)))
+
         return filtered_db
 
 
-    def create_image_list_file(self, image_list_path, mode="relpath"):
+    def create_image_list_file(self, image_list_path, img_list_mode="relpath", append_mode="overwrite"):
         """Create a file containing the list of images in the database.
 
         Parameters:
@@ -510,22 +566,38 @@ class ImageDatabase:
 
         """
 
-        if mode == "abspath":
+        if img_list_mode == "abspath":
             assert self.root_path is not None, "The dataset root path is not set."
 
+        if img_list_mode == "relpath":
+            list_new = self.get_relpaths()
+        elif img_list_mode == "abspath":
+            list_new = [os.path.join(save_db.root_path, img_relpath) for img_relpath in self.get_relpaths()]
+        elif img_list_mode == "name":
+            list_new = [os.path.basename(img_relpath) for img_relpath in self.get_relpaths()]
+        else:
+            assert False, "Invalid mode: " + img_list_mode
+
+        if os.path.isfile(image_list_path):
+            print("- the given image list file already exists")
+            if append_mode in ["keep_new", "keep_old"]:
+                print("  - merging the new image list to the existing one")
+                list_save = merge_img_list(image_list_path, list_new, append_mode)
+            elif append_mode == "overwrite":
+                print("  - overwriting the existing image list file")
+                list_save = list_new
+        else:
+            print("- creating a new image list file")
+            list_save = list_new
+        
+        list_save.sort()
+        
         with open(image_list_path, "wt") as f:
-            for img_relpath in self.db:
-                if mode == "relpath":
+            for img_relpath in list_save:
                     f.write(img_relpath + "\n")
-                elif mode == "abspath":
-                    f.write(os.path.join(self.root_path, img_relpath) + "\n")
-                elif mode == "name":
-                    f.write(os.path.basename(img_relpath) + "\n")
-                else:
-                    assert False, "Invalid mode: " + mode
 
 
-    def copy_images(self, dest_dir, mode="relpath_to_name"):
+    def copy_images(self, dest_dir, mode="relpath_to_name", append_mode="overwrite"):
         """Copy the images in the database to a destination directory.
 
         Parameters:
@@ -534,13 +606,29 @@ class ImageDatabase:
         (all images into single directory and rename them with their original 
         relative paths), "name" (all images into single directory and keep 
         names), or "keep_relpaths" (keep the original directory structure).
+        append_mode (str): The mode of the image saving if the there are 
+        already existing images in the given directory. Can be "keep_new" (keep 
+        the new images in case of overlap), "keep_old" (keep the old images 
+        in case of overlap), or "overwrite" (delete all the old images and save 
+        the new ones).
 
         """
 
         assert self.root_path is not None, "The dataset root path is not set."
         assert os.path.isdir(dest_dir), "The given destination directory does not exist: " + dest_dir
 
+        print("- copying the images to the destination directory")
+        progress = utils.rich_progress_bar()
+        progress.start()
+
+        if append_mode == "overwrite":
+            print("- clearing the destination directory")
+            shutil.rmtree(dest_dir)
+            os.makedirs(dest_dir)
+
+        task = progress.add_task("  ", total=len(self.db))
         for img_relpath in self.db:
+            progress.update(task, advance=1)
             img_name = os.path.basename(img_relpath)
             if mode == "relpath_to_name":
                 img_name = "_".join(img_relpath.split(os.sep))
@@ -564,6 +652,9 @@ class ImageDatabase:
             md5_db = self.db[img_relpath]["md5"]
             md5_src = hashlib.md5(open(img_src_path, 'rb').read()).hexdigest()
 
+            if os.path.isfile(img_dest_path) and append_mode == "keep_old":
+                continue
+            
             shutil.copy2(img_src_path, img_dest_path)
 
             md5_dest = hashlib.md5(open(img_dest_path, 'rb').read()).hexdigest()
@@ -572,6 +663,38 @@ class ImageDatabase:
                 print("WARN: MD5 hash of the source image does not match the one in the database - check integrity of the file: " + img_src_path)
             elif md5_db != md5_dest:
                 print("WARN: MD5 hash of the copied image does not match the one in the database (the source seems OK) - check integrity of the copied file: " + img_dest_path)
+
+        progress.remove_task(task)
+        progress.stop()
+
+
+    def get_subset(self, subset_size, mode, seed):
+        if subset_size == -1:
+            return copy.deepcopy(self)
+        
+        if subset_size >= len(self):
+            print("The subset size is greater or equal to the size of the full retrieved set - returning the whole set.")
+            return copy.deepcopy(self)
+
+        subset_db = ImageDatabase(root_path=self.root_path)
+
+        img_name_list = list(self.db.keys())
+        img_name_list.sort()
+
+        if mode == "random":
+            print("- selecting a random subset of size {} with seed {}".format(subset_size, seed))
+            np.random.seed(seed)
+            subset_indices = np.random.choice(len(img_name_list), subset_size, replace=False)
+            subset_indices.sort()
+        elif mode == "regular":
+            print("- selecting a regular subset of size {}".format(subset_size))
+            subset_indices = np.linspace(0, len(img_name_list) - 1, num=subset_size, dtype=int)
+        subset_img_names = [img_name_list[i] for i in subset_indices]
+
+        for img_name in subset_img_names:
+            subset_db.add_image(img_name, self.db[img_name])
+        
+        return subset_db
 
 
 def compare_metadata(image_value, filter_values, filter_mode, cyclic=False):
@@ -636,6 +759,108 @@ def compare_range_cyclic(value, left_limit = -float("inf"), right_limit = float(
         return left_limit <= value <= right_limit
     else:
         return left_limit <= value or value <= right_limit
+
+
+def str2list(s):
+    """Convert a string to a list of values.
+
+    Parameters:
+    s (str): The input string looking as list.
+
+    Returns:
+    list: The list of values.
+
+    """
+
+    s = s.translate({ord(c): '' for c in "['']"})
+    return [val.strip() for val in s.split(",")]
+
+
+def merge_databases(db_old, db_new, append_mode):
+    """Merge two databases.
+
+    Parameters:
+    db_old (ImageDatabase or str): The old database or the path to the old
+    database file.
+    db_new (ImageDatabase or str): The new database or the path to the new
+    database file.
+    append_mode (str): The mode of the merging. Can be "keep_new" (keep the 
+    new data in case of overlap), "keep_old" (keep the old data in case of 
+    overlap), or "overwrite" (overwrite all of the old data).
+
+    Returns:
+    ImageDatabase: The merged database.
+
+    """
+
+    if isinstance(db_old, str):
+        db_old_path = db_old
+        assert os.path.isfile(db_old_path), "The given database file does not exist: " + db_old
+        db_old = ImageDatabase()
+        db_old.load(db_old_path)
+    
+    if isinstance(db_new, str):
+        db_new_path = db_new
+        assert os.path.isfile(db_new_path), "The given database file does not exist: " + db_new
+        db_new = ImageDatabase()
+        db_new.load(db_new_path)
+
+    if append_mode == "keep_new":
+        merged_db = copy.deepcopy(db_new)
+        for img_name, img_data in db_old.db.items():
+            if img_name not in merged_db:
+                merged_db.add_image(img_name, img_data)
+    elif append_mode == "keep_old":
+        merged_db = copy.deepcopy(db_old)
+        for img_name, img_data in db_new.db.items():
+            if img_name not in merged_db:
+                merged_db.add_image(img_name, img_data)
+    elif append_mode == "overwrite":
+        merged_db = copy.deepcopy(db_new)
+    else:
+        assert False, "Invalid append mode: " + append_mode
+
+    return merged_db
+
+
+def merge_img_list(list_old, list_new, append_mode):
+    """Merge two image lists.
+
+    Parameters:
+    list_old (list or str): The old image list or the path to the old image
+    list file.
+    list_new (list or str): The new image list or the path to the new image
+    list file.
+    append_mode (str): The mode of the file saving if the there are already
+    existing files in the given paths.
+
+    Returns:
+    list: The merged image list.
+    
+    """
+
+    if isinstance(list_old, str):
+        list_old_path = list_old
+        assert os.path.isfile(list_old_path), "The given image list file does not exist: " + list_old
+        with open(list_old_path, "rt") as f:
+            list_old = f.readlines()
+            list_old = [line.strip() for line in list_old]
+    
+    if isinstance(list_new, str):
+        list_new_path = list_new
+        assert os.path.isfile(list_new_path), "The given image list file does not exist: " + list_new
+        with open(list_new_path, "rt") as f:
+            list_new = f.readlines()
+            list_new = [line.strip() for line in list_new]
+
+    if append_mode in ["keep_new", "keep_old"]:
+        merged_list = list(set(list_old + list_new))
+    elif append_mode == "overwrite":
+        merged_list = list_new
+    else:
+        assert False, "Invalid append mode: " + append_mode
+    
+    return merged_list
 
 
 if __name__ == "__main__":
